Placagrama arriba

+++ SIGECO/Models/ValorCatalogogrid.cs
@@ -386,6 +386,9 @@ namespace AspNetMaker2019.Models {
 					ValorCatalogo = this;
 				AddUrl = "ValorCatalogoadd";
 
+				// Table object (Usuario)
+				Usuario = Usuario ?? new _Usuario();
+
 				// Start time
 				StartTime = Environment.TickCount;
 
@@ -394,6 +397,10 @@ namespace AspNetMaker2019.Models {
 
 				// Open connection
 				Conn = Connection; // DN
+
+				// User table object (Usuario)
+				UserTable = UserTable ?? new _Usuario();
+				UserTableConn = UserTableConn ?? GetConnection(UserTable.DbId);
 
 				// List options
 				ListOptions = new ListOptions { TableVar = TableVar };
@@ -599,6 +606,36 @@ namespace AspNetMaker2019.Models {
 				// Header
 				Header(Config.Cache);
 
+				// User profile
+				Profile = new UserProfile();
+
+				// Security
+				Security = new AdvancedSecurity(); // DN
+				bool validRequest = false;
+
+				// Check security for API request
+				if (IsApi() && !Security.IsLoggedIn) {
+					var authResult = await HttpContext.AuthenticateAsync(JwtBearerDefaults.AuthenticationScheme);
+					if (authResult.Succeeded && authResult.Principal.Identity.IsAuthenticated)
+						Security.LoginUser(ClaimValue(ClaimTypes.Name), ClaimValue("userid"), ClaimValue("parentuserid"), ConvertToInt(ClaimValue("userlevelid")));
+				}
+				if (!validRequest) {
+					if (!Security.IsLoggedIn)
+						await Security.AutoLogin();
+					if (Security.IsLoggedIn)
+						Security.TablePermission_Loading();
+					Security.LoadCurrentUserLevel(ProjectID + TableName);
+					if (Security.IsLoggedIn)
+						Security.TablePermission_Loaded();
+					if (!Security.CanList) {
+						Security.SaveLastUrl();
+						FailureMessage = DeniedMessage(); // Set no permission
+						if (IsApi())
+							return new JsonBoolResult(new { success = false, error = DeniedMessage(), version = Config.ProductVersion }, false);
+						return Terminate(GetUrl("Index"));
+					}
+				}
+
 				// Create form object
 				CurrentForm = new HttpForm();
 
@@ -688,6 +725,8 @@ namespace AspNetMaker2019.Models {
 
 				// Build filter
 				filter = "";
+				if (!Security.CanList)
+					filter = "(0=1)"; // Filter all records
 
 				// Restore master/detail filter
 				DbMasterFilter = MasterFilter; // Restore master filter
@@ -742,6 +781,8 @@ namespace AspNetMaker2019.Models {
 
 				// Set no record found message
 				if (Empty(CurrentAction) && TotalRecords == 0) {
+					if (!Security.CanList)
+						WarningMessage = DeniedMessage();
 					if (SearchWhere == "0=101")
 						WarningMessage = Language.Phrase("EnterSearchCriteria");
 					else
@@ -1152,25 +1193,25 @@ namespace AspNetMaker2019.Models {
 				// "view"
 				item = ListOptions.Add("view");
 				item.CssClass = "text-nowrap";
-				item.Visible = true;
+				item.Visible = Security.CanView;
 				item.OnLeft = true;
 
 				// "edit"
 				item = ListOptions.Add("edit");
 				item.CssClass = "text-nowrap";
-				item.Visible = true;
+				item.Visible = Security.CanEdit;
 				item.OnLeft = true;
 
 				// "copy"
 				item = ListOptions.Add("copy");
 				item.CssClass = "text-nowrap";
-				item.Visible = true;
+				item.Visible = Security.CanAdd;
 				item.OnLeft = true;
 
 				// "delete"
 				item = ListOptions.Add("delete");
 				item.CssClass = "text-nowrap";
-				item.Visible = true;
+				item.Visible = Security.CanDelete;
 				item.OnLeft = true;
 
 				// Drop down button for ListOptions
@@ -1226,7 +1267,11 @@ namespace AspNetMaker2019.Models {
 						var options = ListOptions;
 						options.UseButtonGroup = true; // Use button group for grid delete button
 						listOption = options["griddelete"];
-						listOption.Body = "<a class=\"ew-grid-link ew-grid-delete\" title=\"" + HtmlTitle(Language.Phrase("DeleteLink")) + "\" data-caption=\"" + HtmlTitle(Language.Phrase("DeleteLink")) + "\" onclick=\"return ew.deleteGridRow(this, " + RowIndex + ");\">" + Language.Phrase("DeleteLink") + "</a>";
+						if (!Security.CanDelete && IsNumeric(RowIndex) && (RowAction == "" || RowAction == "edit")) { // Do not allow delete existing record
+							listOption.Body = "&nbsp;";
+						} else {
+							listOption.Body = "<a class=\"ew-grid-link ew-grid-delete\" title=\"" + HtmlTitle(Language.Phrase("DeleteLink")) + "\" data-caption=\"" + HtmlTitle(Language.Phrase("DeleteLink")) + "\" onclick=\"return ew.deleteGridRow(this, " + RowIndex + ");\">" + Language.Phrase("DeleteLink") + "</a>";
+						}
 					}
 				}
 				if (CurrentMode == "view") { // View mode
@@ -1234,7 +1279,7 @@ namespace AspNetMaker2019.Models {
 				// "view"
 				listOption = ListOptions["view"];
 				string viewcaption = HtmlTitle(Language.Phrase("ViewLink"));
-				isVisible = true;
+				isVisible = Security.CanView;
 				if (isVisible) {
 					listOption.Body = "<a class=\"ew-row-link ew-view\" title=\"" + viewcaption + "\" data-caption=\"" + viewcaption + "\" href=\"" + HtmlEncode(AppPath(ViewUrl)) + "\">" + Language.Phrase("ViewLink") + "</a>";
 				} else {
@@ -1244,7 +1289,7 @@ namespace AspNetMaker2019.Models {
 				// "edit"
 				listOption = ListOptions["edit"];
 				string editcaption = HtmlTitle(Language.Phrase("EditLink"));
-				isVisible = true;
+				isVisible = Security.CanEdit;
 				if (isVisible) {
 					listOption.Body = "<a class=\"ew-row-link ew-edit\" title=\"" + editcaption + "\" data-caption=\"" + editcaption + "\" href=\"" + HtmlEncode(AppPath(EditUrl)) + "\">" + Language.Phrase("EditLink") + "</a>";
 				} else {
@@ -1254,7 +1299,7 @@ namespace AspNetMaker2019.Models {
 				// "copy"
 				listOption = ListOptions["copy"];
 				string copycaption = HtmlTitle(Language.Phrase("CopyLink"));
-				isVisible = true;
+				isVisible = Security.CanAdd;
 				if (isVisible) {
 					listOption.Body = "<a class=\"ew-row-link ew-copy\" title=\"" + copycaption + "\" data-caption=\"" + copycaption + "\" href=\"" + HtmlEncode(AppPath(CopyUrl)) + "\">" + Language.Phrase("CopyLink") + "</a>";
 				} else {
@@ -1263,7 +1308,7 @@ namespace AspNetMaker2019.Models {
 
 				// "delete"
 				listOption = ListOptions["delete"];
-				isVisible = true;
+				isVisible = Security.CanDelete;
 				if (isVisible)
 					listOption.Body = "<a class=\"ew-row-link ew-delete\"" + "" + " title=\"" + HtmlTitle(Language.Phrase("DeleteLink")) + "\" data-caption=\"" + HtmlTitle(Language.Phrase("DeleteLink")) + "\" href=\"" + HtmlEncode(AppPath(DeleteUrl)) + "\">" + Language.Phrase("DeleteLink") + "</a>";
 				else
@@ -1310,7 +1355,7 @@ namespace AspNetMaker2019.Models {
 					string addcaption = HtmlTitle(Language.Phrase("AddLink"));
 					AddUrl = GetAddUrl();
 					item.Body = "<a class=\"ew-add-edit ew-add\" title=\"" + addcaption + "\" data-caption=\"" + addcaption + "\" href=\"" + HtmlEncode(AppPath(AddUrl)) + "\">" + Language.Phrase("AddLink") + "</a>";
-					item.Visible = (AddUrl != "");
+					item.Visible = (AddUrl != "" && Security.CanAdd);
 				}
 			}
 
@@ -1325,7 +1370,7 @@ namespace AspNetMaker2019.Models {
 						option.UseDropDownButton = false;
 						item = option.Add("addblankrow");
 						item.Body = "<a class=\"ew-add-edit ew-add-blank-row\" title=\"" + HtmlTitle(Language.Phrase("AddBlankRow")) + "\" data-caption=\"" + HtmlTitle(Language.Phrase("AddBlankRow")) + "\" href=\"javascript:void(0);\" onclick=\"ew.addGridRow(this);\">" + Language.Phrase("AddBlankRow") + "</a>";
-						item.Visible = true;
+						item.Visible = Security.CanAdd;
 						ShowOtherOptions = item.Visible;
 					}
 				}
@@ -1766,6 +1811,10 @@ namespace AspNetMaker2019.Models {
 
 			// Delete records (based on current filter)
 			protected async Task<JsonBoolResult> DeleteRows() { // DN
+				if (!Security.CanDelete) {
+					FailureMessage = Language.Phrase("NoDeletePermission"); // No delete permission
+					return JsonBoolResult.FalseResult; // No delete permission
+				}
 				bool result = true;
 				List<Dictionary<string, object>> rsold = null;
 				try {
